requirements.txt
# =============================================================================
# Reply Guy Bot - Production Dependencies (Pinned Versions)
# =============================================================================
# Generated: 2025-11-27
# IMPORTANT: These versions are tested and verified for production.
# Do not use >= constraints in production - use exact pins.

# Twitter/X client
twikit==2.1.0

# HTTP client (required by twikit, supabase)
httpx==0.28.1
requests==2.32.3

# Retry logic
tenacity==8.4.2

# Telegram bot
python-telegram-bot==22.3

# Database (Supabase + dependencies)
supabase==2.16.0
gotrue==2.12.2
postgrest==1.1.1
realtime==2.5.2
storage3==0.12.0
supafunc==0.10.1

# Configuration management
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1

# Security
cryptography==45.0.7

# Browser Automation (CookieBot) - Multiple providers with fallback
# Priority 1: nodriver (best anti-detection, no WebDriver)
nodriver>=0.38

# Priority 2: undetected-chromedriver (Selenium with patches)
undetected-chromedriver>=3.5.5

# Priority 3: playwright (reliable fallback)
playwright==1.49.0
playwright-stealth==1.0.6

# Priority 4: DrissionPage (alternative without WebDriver)
DrissionPage>=4.0.0
